bugfix (sp_mock/frontend): Cleaned up frontend
