--- sp_mock/frontend/src/router/index.ts
@@ -1,30 +1,0 @@
-import { createRouter, createWebHistory } from "vue-router";
-import HomeView from "../views/HomeView.vue";
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: "/",
-      name: "home",
-      component: HomeView,
-    },
-    {
-      path: "/stress-test",
-      name: "stress-test",
-      component: () => import("../views/StressTest.vue"),
-    },
-    // {
-    //   path: '/oauth2/callback',
-    //   name: 'oauth2-callback',
-    //   component: () => import('../views/CallbackView.vue')
-    // },
-    // {
-    //   path: '/poems/:id',
-    //   name: 'poem',
-    //   component: () => import('../views/PoemView.vue')
-    // }
-  ],
-});
-
-export default router;
